style: finished css styling in pet page
